pyproject.toml
[project]
name = "agenticai-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-core>=0.3.0",
    "langchain>=0.3.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "psx>=0.0.1",
    "python-dotenv>=1.0.0",
    "langgraph>=0.2.0",
    "langchain-groq>=0.2.0",
    "langchain-google-genai>=2.0.0",
    "langchain-openai>=0.2.0",
    "plotly>=5.18.0",
    "streamlit>=1.38.0",
    "ta>=0.11.0",
    "pandas-ta>=0.3.14b0",
    "requests>=2.31.0",
]
